% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_clones.R
\name{dedup_clones}
\alias{dedup_clones}
\title{dedup_clones}
\usage{
dedup_clones(snp_pairwise, minimum_dist = 0)
}
\arguments{
\item{snp_pairwise}{Two-dimensional named matrix of pairwise-distances.}

\item{minimum_dist}{Minimum distance to consider two sequences a clone.}
}
\value{
matrix
}
\description{
Deduplicates clones which have a minimum distance between them.
}
\examples{
test = rbind(c(0,1,5,6,5),
             c(2,0,5,6,7),
             c(4,5,0,0,0),
             c(4,5,0,0,0),
             c(4,5,0,0,0),
rownames(test) = paste0(rep('seq', 5), seq(1,5))
colnames(test) = paste0(rep('seq', 5), seq(1,5))
dedup_clones(test)

}
