% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_a_ttest.R
\name{do_a_ttest}
\alias{do_a_ttest}
\title{do_a_ttest}
\usage{
do_a_ttest(variable, our_data, group1, group2)
}
\arguments{
\item{variable}{The column on which to perform the t-test.}

\item{our_data}{The tibble that contains the variable data and a `Treatment` variable with two groups.}

\item{group1}{The first group.}

\item{group2}{The second group.}
}
\value{
double
}
\description{
Performs a paired t-test, omitting NA values, on two groups from a column of a tibble.
}
\examples{
our_data = tibble(variable = c(1,2,3,3,2,7,7,7,8,9), Treatment = c("T1", "T1", "T1", "T1", "T1", "T2", "T2", "T2", "T2", "T2"))
do_a_ttest("variable", our_data, "T1", "T2")

}
